backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
supabase>=2.3.0
python-dotenv>=1.0.0
pydantic[email]>=2.6.0
pydantic-settings>=2.1.0
google-cloud-vision>=3.7.0
httpx>=0.26.0
python-multipart>=0.0.9
fpdf2>=2.8.0
